+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Include component list in platform classification reasons.
 
+* Do not report on unconstrained SDK dependencies.
+
 * New platform component: `build`.
 
 ## 0.10.0